requirements.txt
pyusb>=1.2.1
PyYAML>=6.0
